requirements.txt
dash
dash.dependencies
pandas
plotly.express
plotly.graph_objects
sqlalchemy
gunicorn